+++ crates/controller/src/acl.rs
@@ -23,6 +23,8 @@ pub(crate) async fn check_or_create_kustos_default_permissions(authz: &Authz) ->
         [AccessMethod::Patch, AccessMethod::Get],
     )
     .await?;
+    check_or_create_kustos_role_policy(authz, "user", "/users/me/tariff", [AccessMethod::Get])
+        .await?;
     check_or_create_kustos_role_policy(authz, "user", "/users/find", [AccessMethod::Get]).await?;
     check_or_create_kustos_role_policy(
         authz,